requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
streamlit>=1.30.0
openai>=1.10.0 